write a function that prints anything

+++ 0x10-variadic_functions/3-print_all.c
@@ -26,7 +26,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(list, int));
 					break;
 				case 'f':
-					printf("%s%f", sep, va_arg(list, int));
+					printf("%s%d", sep, va_arg(list, double));
 					break;
 				case 's':
 					str = va_arg(list, char *);
@@ -42,7 +42,6 @@ void print_all(const char * const format, ...)
 			i++;
 		}
 	}
-
 	printf("\n");
 	va_end(list);
 }
